2-COST and start of Poster
To get 2-COST done, I had to include new options in the buggy form that would affect the cost, as per buggy rules. Then I had to create a CSV file to refer to costs for each item and edited app,.py to calculate the cost and return it on completing the form and in the table.