Attempt at 3-MULTI
I've made an attempt at 3-MULTI and got multiple buggies saved to the database.

I switched from using fetchone() to using fetchall(). This unfortunately broke my cost() function and it is now stuck calculating for the first buggy stored in the database for all the buggies. I also did not manage to give the user a way to edit their buggies.